requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.4
joblib==1.2.0
kombu==5.2.4
multidict==6.0.2
nltk==3.7
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
redis==4.3.4
regex==2022.9.13
six==1.16.0
tqdm==4.64.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
XlsxWriter==3.0.3
xlwt==1.3.0
yarl==1.8.1
